django_impl/requirements.txt
asgiref==3.3.0
astroid==2.4.2
colorama==0.4.4
coverage==5.3
django-crispy-forms==1.10.0
django-widget-tweaks==1.4.8
flake8==3.8.4
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.0.1
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.6.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
